+++ src/components/layout/dashboard/AppDashboardLayout.vue
@@ -3,9 +3,11 @@ import { computed } from 'vue'
 
 import AppDashboardLayoutFloating from '@/components/layout/dashboard/AppDashboardLayoutFloating.vue'
 import AppSidebar from '@/components/layout/sidebar/AppSidebar.vue'
+import { useI18n } from '@/composables/i18n/i18n.composable'
 import type { AuthUser } from '@/models/auth-user/authUser.model'
 import { useAuthStore } from '@/stores/auth.store'
 
+const { t } = useI18n()
 const authStore = useAuthStore()
 
 const authUser = computed<AuthUser | null>(() => authStore.authUser)
@@ -30,7 +32,7 @@ function onSignOut(): void {
                 to: {
                   name: 'user-overview',
                 },
-                label: 'Users',
+                label: t('user.label.plural'),
                 icon: 'barChartSquare02',
               },
             ],
